# ironflow-sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountState": {
        "description": "Derived state of an account, for display.",
        "enum": ["ok", "near_limit", "limited", "token_invalid", "never_used"],
        "type": "string"
      },
...
      "AccountWindowResponse": {
        "description": "One usage window of an account.",
        "properties": {
          "model_scope": {"description": "Model family the window applies to, `null` for every model.", "nullable": true, "type": "string"},
          "observed_at": {"description": "When the window was observed.", "format": "date-time", "type": "string"},
          "resets_at": {"description": "When the window resets.", "format": "date-time", "nullable": true, "type": "string"},
          "status": {"$ref": "#/components/schemas/AccountWindowStatus"},
          "utilization": {"description": "Fraction used, `0.0..=1.0`.", "format": "double", "type": "number"},
          "window": {"description": "Window name (`five_hour`, `seven_day`).", "type": "string"}
        },
        "required": ["window", "utilization", "status", "observed_at"],
        "type": "object"
      },
      "AccountWindowStatus": {
        "description": "Status of a usage window, as reported by the provider.",
        "enum": ["allowed", "allowed_warning", "rejected"],
        "type": "string"
      },
...
      "ProviderAccountResponse": {
        "description": "A Provider Account, without its credential.",
        "properties": {
          "alert_threshold": {"description": "Alert threshold.", "format": "double", "type": "number"},
          "auth_failed_at": {"description": "When the provider last rejected the credential.", "format": "date-time", "nullable": true, "type": "string"},
          "created_at": {"description": "Creation timestamp.", "format": "date-time", "type": "string"},
          "created_by": {"description": "Creating user.", "format": "uuid", "nullable": true, "type": "string"},
          "display_name": {"description": "Human-readable name.", "type": "string"},
          "enabled": {"description": "Whether the account may be selected.", "type": "boolean"},
          "expires_at": {"description": "When the credential expires.", "format": "date-time", "type": "string"},
          "id": {"description": "Account ID.", "format": "uuid", "type": "string"},
          "kind": {"description": "Account kind.", "type": "string"},
          "max_concurrency": {"description": "Maximum concurrent steps, `null` for unlimited.", "format": "int32", "minimum": 0, "nullable": true, "type": "integer"},
          "name": {"description": "Unique slug.", "type": "string"},
          "plan": {"description": "Subscription plan.", "nullable": true, "type": "string"},
          "priority": {"description": "Priority, lower is preferred.", "format": "int32", "type": "integer"},
          "state": {"$ref": "#/components/schemas/AccountState"},
          "tags": {"description": "Tags.", "items": {"type": "string"}, "type": "array"},
          "updated_at": {"description": "Last update timestamp.", "format": "date-time", "type": "string"},
          "windows": {"description": "Latest windows.", "items": {"$ref": "#/components/schemas/AccountWindowResponse"}, "type": "array"}
        },
        "required": ["id", "name", "display_name", "kind", "enabled", "priority", "tags", "alert_threshold", "expires_at", "created_at", "updated_at", "state", "windows"],
        "type": "object"
      },
...
      "UpdateProviderAccountRequest": {
        "additionalProperties": false,
        "description": "Request body to update a Provider Account. All fields optional; `name`\nand `kind` are immutable.",
        "properties": {
          "alert_threshold": {"description": "New alert threshold in `(0, 1]`.", "format": "double", "nullable": true, "type": "number"},
          "display_name": {"description": "New display name.", "nullable": true, "type": "string"},
          "enabled": {"description": "Enable or disable.", "nullable": true, "type": "boolean"},
          "expires_at": {"description": "New expiry of the credential.", "format": "date-time", "nullable": true, "type": "string"},
          "max_concurrency": {"description": "New max concurrency; `null` removes the limit.", "format": "int32", "minimum": 0, "nullable": true, "type": "integer"},
          "plan": {"description": "New plan; `null` clears it.", "nullable": true, "type": "string"},
          "priority": {"description": "New priority.", "format": "int32", "nullable": true, "type": "integer"},
          "tags": {"description": "New tags (replaces the list).", "items": {"type": "string"}, "nullable": true, "type": "array"},
          "token": {"description": "Replacement credential, write-only. Checked against the provider.", "nullable": true, "type": "string"}
        },
        "type": "object"
      },
      "ApiKeyResponse": {
        "description": "API key summary (never includes the hash or raw key).",
        "properties": {
          "created_at": {
            "description": "Creation date.",
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "description": "Expiration date.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "description": "API key ID.",
            "format": "uuid",
            "type": "string"
          },
          "is_active": {
            "description": "Whether the key is active.",
            "type": "boolean"
          },
          "key_prefix": {
            "description": "First characters for identification.",
            "type": "string"
          },
          "last_used_at": {
            "description": "Last used date.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "Key name.",
            "type": "string"
          },
          "rate_limit_override": {
            "description": "Per-key rate limit override (requests per minute), if set.",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "scopes": {
            "description": "Granted scopes.",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "scopes",
          "is_active",
          "created_at"
        ],
        "type": "object"
      },
      "ApiKeyScope": {
        "description": "Permission scope for an API key.\n\nEach scope grants access to a specific set of actions.\nA key with no scopes has no permissions.",
        "enum": [
          "workflows_read",
          "runs_read",
          "runs_write",
          "runs_manage",
          "stats_read",
          "accounts_read",
          "accounts_manage",
          "admin"
        ],
        "type": "string"
      },
...
      "CreateApiKeyRequest": {
        "description": "Request body for creating an API key.",
        "properties": {
          "expires_at": {
            "description": "Optional expiration date (ISO 8601).",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "Human-readable name for this key.",
            "type": "string"
          },
          "rate_limit_override": {
            "description": "Optional per-key rate limit override (requests per minute).\nWhen set, the server uses this value instead of the global rate\nlimit for requests authenticated with this key.",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "scopes": {
            "description": "Scopes to grant.",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            },
            "type": "array"
          }
        },
        "required": [
          "name",
          "scopes"
        ],
        "type": "object"
      },
      "CreateApiKeyResponse": {
        "description": "Response returned when creating an API key.\nThe raw key is only shown once.",
        "properties": {
          "created_at": {
            "description": "Creation date.",
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "description": "Expiration date.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "description": "API key ID.",
            "format": "uuid",
            "type": "string"
          },
          "key": {
            "description": "The full raw API key (only returned at creation time).",
            "type": "string"
          },
          "key_prefix": {
            "description": "First characters for identification.",
            "type": "string"
          },
          "name": {
            "description": "Key name.",
            "type": "string"
          },
          "rate_limit_override": {
            "description": "Per-key rate limit override (requests per minute), if set.",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "scopes": {
            "description": "Granted scopes.",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScope"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "key",
          "key_prefix",
          "name",
          "scopes",
          "created_at"
        ],
        "type": "object"
      },
...
      "ScopeEntry": {
        "description": "A scope entry with its machine name and human-readable label.",
        "properties": {
          "description": {
            "description": "Short description.",
            "type": "string"
          },
          "label": {
            "description": "Human-readable label (e.g. \"Runs Read\").",
            "type": "string"
          },
          "value": {
            "description": "Machine-readable scope value (e.g. \"runs_read\").",
            "type": "string"
          }
        },
        "required": [
          "value",
          "label",
          "description"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Thomas Tartrau"
    },
    "description": "REST API for the ironflow workflow engine",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Ironflow REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/provider-accounts/{id}": {
      "delete": {
        "operationId": "delete_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "204": {"description": "Provider account deleted"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Delete a Provider Account and its credential. Admin only.",
        "tags": ["provider-accounts"]
      },
      "get": {
        "operationId": "get_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Get a Provider Account by UUID or name. Admin only.",
        "tags": ["provider-accounts"]
      },
      "patch": {
        "operationId": "update_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProviderAccountRequest"}}}, "description": "Fields to change", "required": true},
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account updated"},
          "400": {"description": "Invalid input"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"},
          "422": {"description": "Credential rejected"},
          "502": {"description": "Provider unreachable"}
        },
        "security": [{"Bearer": []}],
        "summary": "Update a Provider Account. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
...
    "/api/v1/api-keys": {
      "get": {
        "operationId": "list_api_keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of API keys"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all API keys for the authenticated user.",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "# Errors\n\n- 400 if the name is empty or scopes are invalid",
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "description": "API key configuration",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            },
            "description": "API key created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (member trying to assign forbidden scopes)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new API key for the authenticated user.",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/v1/api-keys/scopes": {
      "get": {
        "description": "Admins get all scopes, members only get read-only scopes.",
        "operationId": "available_scopes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScopeEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of available scopes"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Return the list of scopes the current user is allowed to assign to API keys.",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/v1/api-keys/{id}": {
      "delete": {
        "description": "# Errors\n\n- 404 if the key does not exist or belongs to another user",
        "operationId": "delete_api_key",
        "parameters": [
          {
            "description": "API key ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API key not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete an API key owned by the authenticated user.",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
...
}